pyproject.toml
[tool.poetry]
name = "friendlyfl-controller"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
django-tables2 = "^2.6.0"
arrow = "^1.2.3"
celery = "^5.3.1"
redis = "^4.6.0"
django-bootstrap-v5 = "^1.0.11"
django-celery-beat = "^2.5.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
